sbin/cleanerd: use uint32_t for reclaimable block counts
Switch the type of 'reclaimable_blocks' related variables from
'unsigned long' to 'uint32_t' in struct nilfs_cleanerd and struct
nilfs_cldconfig.

Since these variables represent block counts within a single segment,
and the number of blocks per segment is defined as 32-bit in the
on-disk format, using 'uint32_t' is sufficient and appropriate.
This change eliminates the type mismatch with the physical format and
improves consistency.

In addition, modify nilfs_convert_size_to_blocks_per_segment() to
return 'uint32_t' and to clamp the result to the number of blocks per
segment if the calculated value exceeds it.  This prevents invalid
parameter values from being set.

Also, update the syslog format specifiers to use PRIu32.

+++ sbin/cldconfig.h
@@ -90,8 +90,8 @@ struct nilfs_cldconfig {
 	bool cf_use_set_suinfo : 1;
 
 	int cf_log_priority;
-	unsigned long cf_min_reclaimable_blocks;
-	unsigned long cf_mc_min_reclaimable_blocks;
+	uint32_t cf_min_reclaimable_blocks;
+	uint32_t cf_mc_min_reclaimable_blocks;
 };
 
 enum nilfs_selection_policy {
